+++ cloud-yblog-user/cloud-yblog-user1901/src/main/java/com/boot/service/UserAuthorityService.java
@@ -6,5 +6,5 @@ public interface UserAuthorityService {
 
     void changeUserAuthority(UserAuthority UserAuthority);
 
-    int selectAuthorityID(int userid);
+    int selectAuthorityID(long userid);
 }
