+++ cloud-yblog-user/cloud-yblog-user1901/src/main/java/com/boot/service/impl/UserAuthorityServiceImpl.java
@@ -17,7 +17,7 @@ public class UserAuthorityServiceImpl implements UserAuthorityService {
     }
 
     @Override
-    public int selectAuthorityID(int userid) {
+    public int selectAuthorityID(long userid) {
         return userAuthorityMapper.selectAuthorityID(userid);
     }
 }
